Add Favicon with Theme-aware SVG Styling
Introduced a favicon (`logo.svg`) to the project. The `_Layout.cshtml` file has been updated to include the new favicon link. Additionally, the SVG styling has been adjusted with a class (`a`) that adapts its fill color based on the user's preferred color scheme. In light mode, the fill color is `#1d1d1b` (black), while in dark mode, it switches to `#ffffff` (white). This ensures a seamless visual experience for users in different color schemes.